requirements.txt
pandas
psycopg2-binary
sqlalchemy
openpyxl
langchain
langchain-openai
python-dotenv
fastapi
uvicorn